Rename efx.ts to token.ts
